in patientmatching module, added code to standalone GUI to support human review of match results to determine algorithm's performance

+++ src/org/regenstrief/linkage/Record.java
@@ -25,7 +25,7 @@ public class Record {
 	 * context
 	 * 
 	 */
-	public Record(int id, String context) {
+	public Record(long id, String context) {
 		demographics = new Hashtable<String, String>();
 		uid = id;
 		uidContext = context;
